FIX: Standalone images can now have captions.
This was previously broken because the standard settings of the parser would wrap these images into paragraphs. This setting has now been corrected.

With this setup, the image captions now look exactly like blockquotes but are part of the same `<figure>` element as the image itself, which should help screen-readers make sense of this.

Also made standalone images horizontally centered again.