Added RSS for "writings" category.
Explicitly only enable RSS for this category, since we don't care about changes to standalone pages like /about. Again we use the `.Summary` to get a nice lead into the article.

Since writings is the main section of the blog, we link to its RSS feed from the header of every page.