FIX: Style inconsistencies in minimized Hugo HTML output.
This was only visible on the production `--minify` version of the output generated by Hugo. So now, the development version also always minifies the HTML. This collapses all whitespace to a single newline character, so it gets rid of spaces from the source which might influence positioning.

Fixed any styling inconsistencies (that I found) in the minified output as well.