FIX: Ensure content is centered on large screens.
The `mx-auto` we had here didn't work because the `body` has no width set. Instead, we're now using the `container` utility, which sets the width of the container to the minimum width of the current breakpoint. This doesn't change anything for smaller screens but on larger ones, `mx-auto` now works as expected.

Additionally, increased the overall maximum width (relevant on very large screens) to use up more horizontal space than before.