Added `just check` recipe using subfont for local only.
Added as `devDepenency` to ensure its not installed on CI (not needed).

This recipe is meant to be used a local validation only. The tool gives useful output, but does not return a useful return value. It can't really be used in automations as I'd like to use it. For letting it simply "do its thing", it is too opinionated.