Removed `local()` fonts from `@font-family`.
We're trading a little bit of performance for the people that happen to have my fonts installed on their system for consistency.

The deployed and local website will only use the supplied webfonts, so any missing symbols will be obvious while writing the articles. If I used the installed fonts locally, I would not spot these errors on my dev machine (not locally and not on the deployed site).

This gives consistent results and consistent behaviour (and performance) on all machines. It makes things easier to compare.