FIX: Removed `.gitignore` from fonts folder.
This is a funny one. The `.gitignore` file was placed in there to do two things at once:

1. Keep the directory around. This is needed because the `fonttools` command will not _create_ the directory and fail, but if its empty git will not track it
2. Do not check in the generated, subset `woff2` font files

Unfortunately, this `.gitignore` file is copied over by `hugo build` as-is. It is then applied when the CI commits the output files to the `gh-pages` branch. This causes the generated `woff2` files to be ignored here. The result is that these files are missing from the final page deployed via GitHub Pages.

To solve this, I've just added the folder to the `.gitignore` in the root (which won't be copied) and added an empty `.gitkeep` file in the folder to keep it around.