Added filters for each type of pick to list.
The `_index.md` files are required to signal to Hugo that the folder is a section. This makes the section available as a filter (iterating the `.Sections` variable) and automatically generates an extra list for the section itself.

We're using `.RegularPagesRecursive` on the list to show not only the direct entries in the current sections, but also all entries from child sections. This makes the top-level list show _all_ entries while the specific section lists only show entries from their own section.

To always render all sections under "picks", we find the "picks" page and list all sections _it_ knows (sections that are direct descendants).

This was a lot more finicky to set up that I'd expected or hoped...