Added RSS feeds for Picks.
This includes filtered sub-feeds for specific types of picks. This is basically "free" with how the page loading works.

I had to find a workaround to render the `category`, which was supposed to just be the name of the `.Section`. Unfortunately, this resolves not to the _direct_ parent, but to the top-most one. Not sure why. So instead, I go up the `.Ancestors` list by _one_ step and then use that.