Automatically link RSS feeds for all parent sections.
Previously, the feed was hardcoded. Now, its auto generated based on the current Page the user is on.

If it's a normal page, the page itself and all its ancestors are checked if they have an RSS output format - every candidate that does is added. This way, for example a specific Pick page will add both the general "All picks" feed and the specific feed of its type.

If it's the home page, we pick both the "Writings" list and the "All Picks" list manually and add the feeds to the list. This is to advertise the most prominent feeds on the home page directly.