FIX: Ensure the main menu does not scroll page horizontally.
With the new "picks" entry, the menu got too wide. I have reduced the text size on mobile to cram more stuff into a single line.

Also refactored the header markup to use flex and adaptable margins to align text and image vertically/horizontally. This gets rid of the "float" stuff which worked but broke in corner cases.