Set the meta description for picks lists.
This was mistakenly rendering the actual content as the summary, which isn't what we want. It mostly shows up when linking to the page in rich link previews, so I didn't realize it earlier.

I have duplicated the content into each `_index.md` file and set the `summary` property specifically. While this isn't perfect, I'm also now rendering it directly on the list so it should be less easy to forget setting it on new pick types. Time will tell...